requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pdfplumber==0.10.3
openai==1.10.0
ragas==0.0.22
datasets==2.16.1
pandas==2.2.0
langchain-openai==0.0.5
alembic==1.13.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
ragas